requirements.txt
dbt-snowflake
snowflake-connector-python
python-dotenv
